src/requirements.txt
streamlit
openai
python-dotenv
azure.ai.inference